Refactored gzip format support into lib
