+++ tests/lib/gzip.py
@@ -1,0 +1,15 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+"""Tests for gzip compressed stream file."""
+
+from __future__ import unicode_literals
+
+import unittest
+
+
+# TODO: add tests for _GzipDecompressorState
+# TODO: add tests for GzipMember
+
+
+if __name__ == '__main__':
+  unittest.main()
